requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure
azure-functions
azure-identity 
azure-keyvault

# Utility
beautifulsoup4
requests
openai